+++ invocations/packaging/release.py
@@ -275,16 +275,22 @@ def all_(c, dry_run=False):
     .. versionchanged:: 2.1
         Added the ``dry_run`` flag.
     """
-    prepare(c)
+    prepare(c, dry_run=dry_run)
     publish(c, dry_run=dry_run)
     push(c, dry_run=dry_run)
 
 
 @task
-def prepare(c):
+def prepare(c, dry_run=False):
     """
     Edit changelog & version, git commit, and git tag, to set up for release.
-    
+
+    :param bool dry_run:
+        Whether to take any actual actions or just say what might occur.
+        Default: ``False``.
+
+    .. versionchanged:: 2.1
+        Added the ``dry_run`` parameter.
     .. versionchanged:: 2.1
         Generate annotated git tags instead of lightweight ones.
     """
@@ -296,8 +302,9 @@ def prepare(c):
     # transmitted from status() into here...
     actions, state = status(c)
     # TODO: unless nothing-to-do in which case just say that & exit 0
-    if not confirm("Take the above actions?"):
-        raise Exit("Aborting.")
+    if not dry_run:
+        if not confirm("Take the above actions?"):
+            raise Exit("Aborting.")
 
     # TODO: factor out what it means to edit a file:
     # - $EDITOR or explicit expansion of it in case no shell involved
@@ -309,7 +316,7 @@ def prepare(c):
         # TODO: identify top of list and inject a ready-made line? Requires vim
         # assumption...GREAT opportunity for class/method based tasks!
         cmd = "$EDITOR {.packaging.changelog_file}".format(c)
-        c.run(cmd, pty=True, hide=False)
+        c.run(cmd, pty=True, hide=False, dry=dry_run)
     # Version file!
     if actions.version == VersionFile.NEEDS_BUMP:
         version_file = os.path.join(
@@ -317,7 +324,7 @@ def prepare(c):
             c.packaging.get("version_module", "_version") + ".py",
         )
         cmd = "$EDITOR {}".format(version_file)
-        c.run(cmd, pty=True, hide=False)
+        c.run(cmd, pty=True, hide=False, dry=dry_run)
     if actions.tag == Tag.NEEDS_CUTTING:
         # Commit, if necessary, so the tag includes everything.
         # NOTE: this strips out untracked files. effort.
@@ -326,14 +333,15 @@ def prepare(c):
             c.run(
                 'git commit -am "Cut {}"'.format(state.expected_version),
                 hide=False,
+                dry=dry_run,
             )
         # Tag!
         c.run(
             'git tag -a {} -m ""'.format(state.expected_version),
             hide=False,
+            dry=dry_run,
+            echo=True,
         )
-        # TODO: print something to clarify/confirm tag was cut, if not just
-        # adding echo=True to above
 
 
 def _release_line(c):
@@ -699,6 +707,8 @@ def publish(
     """
     # Don't hide by default, this step likes to be verbose most of the time.
     c.config.run.hide = False
+    # Including echoing!
+    c.config.run.echo = True
     # Config hooks
     # TODO: this pattern is too widespread. Really needs something in probably
     # Executor that automatically does this on our behalf for any kwargs we
@@ -750,7 +760,8 @@ def upload(c, directory, index=None, sign=False, dry_run=False):
         Whether to sign the built archive(s) via GPG.
 
     :param bool dry_run:
-        Skip actual publication step if ``True``.
+        Skip actual publication step (and dry-run actions like signing) if
+        ``True``.
 
         This also prevents cleanup of the temporary build/dist directories, so
         you can examine the build artifacts.
@@ -769,7 +780,8 @@ def upload(c, directory, index=None, sign=False, dry_run=False):
     # doesn't allow you to dry-run or upload manually when API is borked...
     if sign:
         prompt = "Please enter GPG passphrase for signing: "
-        input_ = StringIO(getpass.getpass(prompt) + "\n")
+        passphrase = "" if dry_run else getpass.getpass(prompt)
+        input_ = StringIO(passphrase + "\n")
         gpg_bin = find_gpg(c)
         if not gpg_bin:
             raise Exit(
@@ -780,7 +792,7 @@ def upload(c, directory, index=None, sign=False, dry_run=False):
             cmd = "{} --detach-sign -a --passphrase-fd 0 {{}}".format(
                 gpg_bin
             )  # noqa
-            c.run(cmd.format(archive), in_stream=input_)
+            c.run(cmd.format(archive), in_stream=input_, dry=dry_run)
             input_.seek(0)  # So it can be replayed by subsequent iterations
     # Upload
     parts = ["twine", "upload"]
@@ -805,7 +817,7 @@ def push(c, dry_run=False):
     Push current branch and tags to default Git remote.
     """
     kwargs = dict(echo=True) if dry_run else dict()
-    opts = " --dry-run" if dry_run else ""
+    opts = " --dry-run --no-verify" if dry_run else ""
     c.run("git push --follow-tags{}".format(opts), **kwargs)